* refactor: ProductDetail 분리
